+++ tests/test_custom_policy.py
@@ -13,39 +13,53 @@ from stable_baselines.sac.policies import FeedForwardPolicy as SACPolicy
 
 N_TRIALS = 100
 
+
 class CustomCommonPolicy(FeedForwardPolicy):
     def __init__(self, *args, **kwargs):
+        # Default value
+        if 'net_arch' not in kwargs:
+            kwargs['net_arch'] = [8, dict(vf=[8, 8], pi=[8, 8])]
         super(CustomCommonPolicy, self).__init__(*args, **kwargs,
-                                           net_arch=[8, dict(vf=[8, 8], pi=[8, 8])],
-                                           feature_extraction="mlp")
+                                                 feature_extraction="mlp")
+
 
 class CustomDQNPolicy(DQNPolicy):
     def __init__(self, *args, **kwargs):
+        # Default value
+        if 'layers' not in kwargs:
+            kwargs['layers'] = [8, 8]
         super(CustomDQNPolicy, self).__init__(*args, **kwargs,
-                                           layers=[8, 8],
-                                           feature_extraction="mlp")
+                                              feature_extraction="mlp")
+
 
 class CustomDDPGPolicy(DDPGPolicy):
     def __init__(self, *args, **kwargs):
+        # Default value
+        if 'layers' not in kwargs:
+            kwargs['layers'] = [8, 8]
         super(CustomDDPGPolicy, self).__init__(*args, **kwargs,
-                                           layers=[8, 8],
-                                           feature_extraction="mlp")
+                                               feature_extraction="mlp")
+
 
 class CustomSACPolicy(SACPolicy):
     def __init__(self, *args, **kwargs):
+        # Default value
+        if 'layers' not in kwargs:
+            kwargs['layers'] = [8, 8]
         super(CustomSACPolicy, self).__init__(*args, **kwargs,
-                                           layers=[8, 8],
-                                           feature_extraction="mlp")
+                                              feature_extraction="mlp")
 
+
+# MODEL_CLASS, POLICY_CLASS, POLICY_KWARGS
 MODEL_DICT = {
-    'a2c': (A2C, CustomCommonPolicy, dict(act_fun=tf.nn.relu)),
+    'a2c': (A2C, CustomCommonPolicy, dict(act_fun=tf.nn.relu, net_arch=[12, dict(vf=[16], pi=[8])])),
     'acer': (ACER, CustomCommonPolicy, dict(act_fun=tf.nn.relu)),
     'acktr': (ACKTR, CustomCommonPolicy, dict(act_fun=tf.nn.relu)),
-    'dqn': (DQN, CustomDQNPolicy, dict()),
-    'ddpg': (DDPG, CustomDDPGPolicy, dict()),
-    'ppo1': (PPO1, CustomCommonPolicy, dict(act_fun=tf.nn.relu)),
-    'ppo2': (PPO2, CustomCommonPolicy, dict(act_fun=tf.nn.relu)),
-    'sac': (SAC, CustomSACPolicy, dict()),
+    'dqn': (DQN, CustomDQNPolicy, dict(layers=[4, 4], dueling=False)),
+    'ddpg': (DDPG, CustomDDPGPolicy, dict(layers=[16, 16], layer_norm=False)),
+    'ppo1': (PPO1, CustomCommonPolicy, dict(act_fun=tf.nn.relu, net_arch=[8, 4])),
+    'ppo2': (PPO2, CustomCommonPolicy, dict(act_fun=tf.nn.relu, net_arch=[4, 4])),
+    'sac': (SAC, CustomSACPolicy, dict(layers=[16, 16])),
     'trpo': (TRPO, CustomCommonPolicy, dict(act_fun=tf.nn.relu)),
 }
 
@@ -54,7 +68,7 @@ MODEL_DICT = {
 def test_custom_policy(model_name):
     """
     Test if the algorithm (with a custom policy) can be loaded and saved without any issues.
-    :param model_class: (BaseRLModel) A RL model
+    :param model_name: (str) A RL model
     """
 
     try:
@@ -78,7 +92,7 @@ def test_custom_policy(model_name):
         model.save("./test_model")
         del model, env
         # loading
-        model = model_class.load("./test_model", policy=policy)
+        _ = model_class.load("./test_model", policy=policy)
 
     finally:
         if os.path.exists("./test_model"):
@@ -89,12 +103,16 @@ def test_custom_policy(model_name):
 def test_custom_policy_kwargs(model_name):
     """
     Test if the algorithm (with a custom policy) can be loaded and saved without any issues.
-    :param model_class: (BaseRLModel) A RL model
+    :param model_name: (str) A RL model
     """
 
     try:
         model_class, policy, policy_kwargs = MODEL_DICT[model_name]
         env = 'MountainCarContinuous-v0' if model_name in ['ddpg', 'sac'] else 'CartPole-v1'
+
+        # Should raise an error when a wrong keyword is passed
+        with pytest.raises(ValueError):
+            model_class(policy, env, policy_kwargs=dict(this_throws_error='maybe'))
 
         # create and train
         model = model_class(policy, env, policy_kwargs=policy_kwargs)
@@ -117,9 +135,9 @@ def test_custom_policy_kwargs(model_name):
         model.learn(total_timesteps=100, seed=0)
         del model
 
-        # Load wit different wrong policy_kwargs
+        # Load with different wrong policy_kwargs
         with pytest.raises(ValueError):
-            model = model_class.load("./test_model", policy=policy, env=env, policy_kwargs=dict(wrong="kwargs"))
+            _ = model_class.load("./test_model", policy=policy, env=env, policy_kwargs=dict(wrong="kwargs"))
 
     finally:
         if os.path.exists("./test_model"):