+++ tests/test_vec_envs.py
@@ -23,7 +23,7 @@ class CustomGymEnv(gym.Env):
         return self.state
 
     def step(self, action):
-        reward = self._get_reward(action)
+        reward = 1
         self._choose_next_state()
         self.current_step += 1
         done = self.current_step >= self.ep_length