+++ tests/test_vec_normalize.py
@@ -56,4 +56,5 @@ def test_mpi_moments():
     test running mean std function
     """
     subprocess.check_call(['mpirun', '--allow-run-as-root', '-np', '3', 'python', '-c',
-                           'from stable_baselines.common.mpi_moments import _helper_runningmeanstd; _helper_runningmeanstd()'])
+                           'from stable_baselines.common.mpi_moments '
+                           'import _helper_runningmeanstd; _helper_runningmeanstd()'])